// src/api/generated/model/mockInvestmentOrderHistoryItemResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * 주머니(Jumoney) API 문서
 * 주머니의 백엔드 API 명세서입니다.
 * OpenAPI spec version: 1.0.0
 */

export interface MockInvestmentOrderHistoryItemResponse {
  orderId?: number;
  orderType?: string;
  stockCode?: string;
  stockName?: string;
  executionPrice?: number;
  quantity?: number;
  totalExecutionAmount?: number;
  executedAt?: string;
}